packages/transformation/pyproject.toml
[project]
name = "transformation"
version = "0.0.0"
description = "DataPlatform internal transformation lib"
readme = "README.md"
authors = [{ name = "javigoncompte" }]
requires-python = ">=3.12"
dynamic = ["una"]
dependencies = [
    "client",
    "core",
    "ruamel-yaml",
    "paramiko",
    "databricks-sdk~=0.65.0",
]

[build-system]
requires = ["hatchling", "hatch-una"]
build-backend = "hatchling.build"

[dependency-groups]
dev = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.hooks.una-build]
[tool.hatch.metadata.hooks.una-meta]

[tool.uv.sources]
core = { workspace = true }
client = { workspace = true }
